Moved interfaces into Interface folder
