+++ Services/Interfaces/IAuthorsService.cs
@@ -1,7 +1,7 @@
 using Newspaper.DTOs;
 using Newspaper.Entities;
 
-namespace Newspaper.Services;
+namespace Newspaper.Services.Interfaces;
 
 public interface IAuthorsService
 {
